+++ README.md
@@ -4,6 +4,11 @@
 
 This project is aimed to analyse the emotional impact of movies conveyed in critical reviews from a large movie dataset.
 
+There will be three steps:
+1.  Acquiring Datas (scraping, API call)
+1. Treatment on datas before loading
+2. Loading on db
+
 ## Requirements
 
 - Git
